// apps/web/src/gen/types/AssignmentStatus.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const assignmentStatusEnum = {
	assigned: "assigned",
	accepted: "accepted",
	arrived: "arrived",
	completed: "completed",
	cancelled: "cancelled",
} as const;

export type AssignmentStatusEnumKey =
	(typeof assignmentStatusEnum)[keyof typeof assignmentStatusEnum];

export type AssignmentStatus = AssignmentStatusEnumKey;