// apps/web/src/gen/types/OccurrencePriority.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const occurrencePriorityEnum = {
	low: "low",
	medium: "medium",
	high: "high",
	critical: "critical",
} as const;

export type OccurrencePriorityEnumKey =
	(typeof occurrencePriorityEnum)[keyof typeof occurrencePriorityEnum];

export type OccurrencePriority = OccurrencePriorityEnumKey;