// apps/web/src/gen/types/OccurrenceStatus.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const occurrenceStatusEnum = {
	open: "open",
	dispatched: "dispatched",
	in_progress: "in_progress",
	resolved: "resolved",
	cancelled: "cancelled",
} as const;

export type OccurrenceStatusEnumKey =
	(typeof occurrenceStatusEnum)[keyof typeof occurrenceStatusEnum];

export type OccurrenceStatus = OccurrenceStatusEnumKey;