// apps/web/src/gen/types/VehicleStatus.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const vehicleStatusEnum = {
	available: "available",
	dispatched: "dispatched",
	maintenance: "maintenance",
	inactive: "inactive",
} as const;

export type VehicleStatusEnumKey =
	(typeof vehicleStatusEnum)[keyof typeof vehicleStatusEnum];

export type VehicleStatus = VehicleStatusEnumKey;